+++ .env
@@ -1,0 +1,4 @@
+REACT_APP_GET_FORMS=http://localhost:5002/form/
+GET_CATEGORIES=http://localhost:5002/GetCategories
+GET_QUESTION=http://localhost:5002/GetQuestions
+GET_USER_DATA=http://localhost:5002/userData/
